move shared to roles under shared folder, added auth persistence intended to be for use with both web apps
