+++ SunDoeCoffeeShop/src/Shared/Roles/RoleConverter.cs
@@ -11,7 +11,7 @@
 // WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-namespace Shared;
+namespace SunDoeCoffeeShop.Shared.Roles;
 
 public static class RoleConverter
 {
