+++ SunDoeCoffeeShop/src/Shared/Roles/SunDoeCoffeeShop.Shared.Roles.csproj
